requirements.txt
gunicorn    # necesario para desplegar/ejecutar en Heroku
Flask
pymongo
dnspython
python-dotenv
google-cloud-texttospeech
flickrapi